+++ app/common/services/datapackage/dataservice.js
@@ -1,7 +1,9 @@
 // import ono from 'ono';
-import {toJS, extendObservable, asFlat, isObservable, autorun} from 'mobx';
+import {toJS, autorun, isObservable} from 'mobx';
 
-import dp from './datapackage';
+import {default as dp, Package, makeResource} from './datapackage';
+
+const DEBUG = !System.production;
 
 // window.process = process;  // annoying
 run.$inject = ['$http'];
@@ -16,7 +18,7 @@ export function run ($http) {
 
 dataservice.$inject = ['$log', 'growl'];
 export function dataservice ($log, growl) {
-  if (!System.production) {
+  if (DEBUG) {
     const processResource = dp.processor.resource.bind(dp.processor);
     dp.processor.resource = resource => {
       $log.debug('Processing', resource.name);
@@ -36,29 +38,46 @@ export function dataservice ($log, growl) {
     mime: dp.normalize.mime,
     translators: dp.processor.translators,
     index: dp.Normalizer.index,
-    warn: (err, opts) => {
+    warn (err, opts) {
       $log.warn(err);
-      growl.warning(err.message || String(err), opts);
+      growl.warning(err.message || String(err), opts); // TODO: make optional
     },
-    error: (err, opts) => {
+    error (err, opts) {
       $log.error(err);
       growl.error(err.message || String(err), opts);
       return Promise.reject(err);
     },
-    loadPackage: async(url, growl = false) => {
-      let datapackage = await __loadPackageJson(url, growl);
+    async loadPackage (datapackage) { // TODO: errors
+      try {
+        return await makePackage(datapackage).load();
+      } catch (err) {
+        const msg = (err.status) ? `${err.statusText}; ${err.data}` : String(err);
+        return dataservice.error(msg, {title: 'Error loading DataPackage'});
+      }
+
+      /* let datapackage = await __loadPackageJson(url, growl);
       datapackage = await __loadPackageResources(datapackage, growl);
-      return dataservice.makePackageObservable(datapackage);
+      return dataservice.makePackageObservable(datapackage); */
       // return __processPackageResources(datapackage, growl);
     },
-    processPackage: async (url, datapackage, growl = false) => {
-      datapackage = __normalizePackage(url, datapackage);
+    async processPackage (url, datapackage) { // TODO: errors
+      return await makePackage(url).update(datapackage);
+
+      /* datapackage = __normalizePackage(url, datapackage);
       datapackage = __normalizePackageResourcesAndSchema(datapackage);
       await __loadPackageResources(datapackage, growl);
-      return dataservice.makePackageObservable(datapackage);
+      return dataservice.makePackageObservable(datapackage); */
       // return __processPackageResources(datapackage, growl);
     },
-    loadResource: async (datapackage, res) => {
+    async loadResource (datapackage, res) {  // TODO: move to datapackage store action
+      try {
+        return await makeResource(datapackage, res).load();
+      } catch (err) {
+        const msg = (err.status) ? `${err.statusText}; ${err.data}` : String(err);
+        return dataservice.error(msg, {title: 'Error loading resource'});
+      }
+
+      /*
       try {
         res = dp.normalize.resource(datapackage, res);
         if (!res.data) {
@@ -68,14 +87,15 @@ export function dataservice ($log, growl) {
         const msg = (err.status) ? `${err.statusText}; ${err.data}` : String(err);
         return dataservice.error(msg, {title: 'Error loading resource'});
       }
-      return dataservice.processResource(res);
+      return dataservice.makeResourceObservable(res);
+      // return dataservice.processResource(res); */
     },
-    normalizeResource: (p, r) => {
+    normalizeResource (p, r) { // deprecated
       return Object.assign(r, dp.normalize.resource(p, r));
     },
-    processResource: __processResource,
-    normalizePackage: __normalizePackage,
-    reloadResource: async resource => {
+    processResource: __processResource,  // deprecated
+    normalizePackage: __normalizePackage,  // deprecated
+    async reloadResource (resource) {  // deprecated
       try {
         Object.assign(resource, await dp.loader.resource(resource));
         return dp.processResource(resource);
@@ -84,7 +104,8 @@ export function dataservice ($log, growl) {
         return resource;  // reject?
       }
     },
-    makeResourceObservable: resource => {
+    makeResourceObservable: makeResource, // deprecated
+    /* resource => {
       if (isObservable(resource)) {
         return resource;
       }
@@ -93,58 +114,71 @@ export function dataservice ($log, growl) {
         extendObservable(resource, {
           content: resource.content,
           mediatype: resource.mediatype,
-          name: resource.name,
-          get __data__ () {
-            return dp.processor.resource(toJS(resource));
-          },
-          get data () {
-            return resource.__data__.data;
-          },
-          get $error () {
-            return resource.__data__.$error;
-          },
-          get errors () {
-            return resource.__data__.errors;
-          }
+          name: resource.name
         });
       }
 
       autorun(() => {
-        if (resource.$error) {
-          resource.$error.message = resource.name;
-          dataservice.warn(resource.$error, {title: 'Error processing resource'});
-          resource.$valid = false;
-        }
-        $log.debug(`${resource.name} updated`);
+        __processResource(resource);
       });
 
       return resource;
-    },
-    makePackageObservable: datapackage => {
+    },*/
+    makePackageObservable: makePackage // deprecated
+    /* datapackage => {
       if (isObservable(datapackage)) {
         return datapackage;
       }
 
-      return extendObservable(datapackage, {
-        resources: asFlat(datapackage.resources.map(dataservice.makeResourceObservable)),
-        get resourcesByName () {
-          return dp.Normalizer.index(datapackage);
-        },
-        get $resourcesByName () {
-          return datapackage.resourcesByName();
-        }
+      datapackage = extendObservable(datapackage, {
+        resources: asFlat(datapackage.resources.map(dataservice.makeResourceObservable))
       });
-    }
+
+      autorun(() => {
+        datapackage.$resourcesByName = datapackage.resourcesByName = dp.Normalizer.index(datapackage);
+        $log.debug(`${datapackage.name} re-indexed`);
+      });
+
+      return datapackage;
+    }*/
   };
 
-  function __processResource (res) {
-    dp.processResource(res);
-    if (res.$error) {
-      res.$error.message = res.name;
-      dataservice.warn(res.$error, {title: 'Error processing resource'});
-      return Object.assign(res, {$valid: false});
+  function __processResource (resource) { // deprecated
+    // get process results
+    let processed = toJS(Object.assign({}, resource, {
+      $error: null,
+      errors: [],
+      data: []
+    }));
+    try {
+      processed = dp.processor.resource(processed);
+      if (processed.errors && processed.errors.length > 0) {
+        processed.$error = new Error(`Errors processing resource ${processed.name}`);
+      }
+    } catch (err) {
+      processed.$error = err;
+      processed.errors = processed.errors || [];
+      processed.errors.shift({
+        code: 'Parsing',
+        type: err.name,
+        message: `Parsing error: ${err.message}`
+      });
     }
-    return res;
+
+    // update resource
+    if (processed.$error || processed.errors.length > 0) {
+      resource.errors = processed.errors;
+      resource.$error = processed.$error;
+      resource.$error.message = resource.name;
+      resource.$valid = false;
+      dataservice.warn(resource.$error, {title: 'Error processing resource'});
+    } else {
+      resource.errors = [];
+      resource.$error = null;
+      resource.data = processed.data;
+      resource.$lastValidContent = toJS(resource.content);
+    }
+    return resource;
   }
 
   /* function __processPackageResources (datapackage, growl = false) {
@@ -164,7 +198,7 @@ export function dataservice ($log, growl) {
     }
   } */
 
-  async function __loadPackageResources (datapackage, growl = false) {
+  /* async function __loadPackageResources (datapackage, growl = false) {
     try {
       return await dp.loadResources(datapackage);
     } catch (err) {
@@ -177,9 +211,9 @@ export function dataservice ($log, growl) {
       }
       return Promise.reject(msg);
     }
-  }
+  } */
 
-  function __normalizePackage (url, datapackage) {
+  function __normalizePackage (url, datapackage) { // deprecated
     if (arguments.length === 2) {
       Object.assign(datapackage, {url});
     } else {
@@ -188,12 +222,12 @@ export function dataservice ($log, growl) {
     return dp.normalizePackage(datapackage);
   }
 
-  function __normalizePackageResourcesAndSchema (datapackage) {
+  /* function __normalizePackageResourcesAndSchema (datapackage) {
     datapackage = dp.normalizeResources(datapackage);
     return dp.normalizeSchemas(datapackage);
-  }
+  } */
 
-  async function __loadPackageJson (url, growl = false) {
+  /* async function __loadPackageJson (url, growl = false) {
     try {
       let datapackage = await dp.loader.datapackage(url);
       datapackage = __normalizePackage(datapackage);
@@ -208,6 +242,24 @@ export function dataservice ($log, growl) {
       }
       return Promise.reject(msg);
     }
+  } */
+
+  function makePackage (datapackage) {
+    if (isObservable(datapackage)) {
+      return datapackage;
+    }
+    datapackage = new Package(datapackage);
+
+    autorun(() => {  // TODO: make optional
+      datapackage.resources.forEach(resource => {
+        if (resource.errors && resource.errors.length > 0) {
+          return dataservice.error(`Errors processing resource ${resource.name}`,
+            {title: `Error processing DataPackage`});
+        }
+      });
+    });
+
+    return datapackage;
   }
 
   return dataservice;